* 'master' of https://github.com/hazelcast/hazelcast: (598 commits)
  Fixed a compiletime issue on the ClientMapProxy
  test and serialization bugfix
  Fixed wrong StoreFactory
  executeOnKeys(keys,entryprocessor) introduced on IMap. With this feature you can run entry processor on a set of keys
  Fixed done assignment
  Codestyle :-)
  Fixed missing keystore algorithm on IBM JDKs
  fixed missing done assignment
  test for issue #1404
  set default queue capacity to Integer.MAX_VALUE
  merging pull request #1377 Optimizations - MapGetAllOperation doesn't have to be always invoked on all partitions
  Removing near cache on destroying MapProxy
  added test
  fixes #1401, fixed loading of polled item from store. clear unnecessary warning
  Implemented ExecutorService::invokeAll(Collection, long, TimeUnit) and fixed an unexpected CancellationException throw in DistributedExecutorService
  removed ignore comment
  add test for issue 254 Vtopo_TTL_Replication_Issue254
  Added functionality for controlling the executor to use from an operation
  Added copyrights.
  Marked ReplicaSyncResponse with UrgentSystemOperation.
  Added urgency to the IO system so that urgent packets are send with a higher precedence than regular packets.
  client io Test Executor Pool Size Low Simulating A high amount of listeners and Async Opps @Ignore("Issue #1391")
  client io Test Executor Pool Size Low Simulating A high amount of listeners and Async Opps
  Deleted AbstractNodeAware since unused
  Removed ClientInvocation since unused
  Improved reliability by adding priority execution for System Operations.
  client io Test adding more listeners that executor threads
  handle NULL_OBJECT inside NearCache
  client io Test
  client IO stress test
  Renamed OperationServiceImpl to InternalOperationService
  Typo corrected: `Facyory` in Section 15.
  containsKey() checks near cache first
  fixed missing argument when creating transactionContext
  fixed #850 - DelegatingFuture.isDone always returns false until get is called
  Added nearcache testcase
  @override for implementations of IMap interface
  Fixed put to nearcache
  Added record access for lastAccess time in statistics
  Minor javadoc improvements + added missing copyright section
  Removed MapProxy interface since not used
  Refactored the code so that a new OperationService can be injected.
  Invocation is now an implementation detail
  unused method stub removed
  created multi issues and marked function with number
  adding ignore issue #1368 for failing tests
  added comment to //Fail comment to failing tests
  Wan Replication Tests update to higher hits merge policy test
  Wan Replication Tests update to higher hits merge policy test
  version info included to pdf file name
  ...

+++ hazelcast/src/main/java/com/hazelcast/map/MapStoreWrapper.java
@@ -19,8 +19,10 @@ package com.hazelcast.map;
 import com.hazelcast.core.MapLoader;
 import com.hazelcast.core.MapLoaderLifecycleSupport;
 import com.hazelcast.core.MapStore;
+import com.hazelcast.core.PostProcessingMapStore;
 
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -48,6 +50,11 @@ public class MapStoreWrapper implements MapStore {
         this.mapLoader = loader;
         this.mapStore = store;
         this.enabled.set(enabled);
+    }
+
+
+    public MapStore getMapStore() {
+        return mapStore;
     }
 
     public void enable() {
@@ -95,6 +102,9 @@ public class MapStoreWrapper implements MapStore {
     }
 
     public void deleteAll(Collection keys) {
+        if( keys == null || keys.isEmpty() ){
+            return;
+        }
         if (isMapStore() && enabled.get()) {
             mapStore.deleteAll(keys);
         }
@@ -115,10 +125,17 @@ public class MapStoreWrapper implements MapStore {
     }
 
     public Map loadAll(Collection keys) {
+        if( keys == null || keys.isEmpty() ){
+            return Collections.EMPTY_MAP;
+        }
         if (isMapLoader() && enabled.get()) {
             return mapLoader.loadAll(keys);
         }
         return null;
+    }
+
+    public boolean isPostProcessingMapStore() {
+        return isMapStore() && mapStore instanceof PostProcessingMapStore;
     }
 
     @Override
